test: getLambdaFunctionContents.spec.ts
